+++ lib/Buzz/Client/Curl.php
@@ -2,16 +2,41 @@
 
 namespace Buzz\Client;
 
+use Buzz\Converter\RequestConverter;
+use Buzz\Converter\ResponseConverter;
 use Buzz\Exception\RequestException;
 use Buzz\Message\MessageInterface;
 use Buzz\Message\RequestInterface;
 use Buzz\Exception\LogicException;
+use GuzzleHttp\Psr7\Response;
+use Psr\Http\Message\RequestInterface as PSR7RequestInterface;
+use Psr\Http\Message\ResponseInterface;
 
 class Curl extends AbstractCurl
 {
     private $lastCurl;
 
+    /**
+     * @param RequestInterface $request
+     * @param MessageInterface $response
+     * @param array $options
+     *
+     * @deprecated Will be removed in 1.0. Use sendRequest instead.
+     */
     public function send(RequestInterface $request, MessageInterface $response, array $options = array())
+    {
+        @trigger_error('Curl::send() is deprecated. Use Curl::sendRequest instead.', E_USER_DEPRECATED);
+        $request = RequestConverter::psr7($request);
+        ResponseConverter::copy(ResponseConverter::buzz($this->sendRequest($request, $options)), $response);
+    }
+
+    /**
+     * @param PSR7RequestInterface $request
+     * @param array $options
+     *
+     * @return ResponseInterface
+     */
+    public function sendRequest(PSR7RequestInterface $request, $options = [])
     {
         if (is_resource($this->lastCurl)) {
             curl_close($this->lastCurl);
@@ -32,7 +57,7 @@ class Curl extends AbstractCurl
             throw $e;
         }
 
-        static::populateResponse($this->lastCurl, $data, $response);
+        return $this->createResponse($this->lastCurl, $data);
     }
 
     /**