+++ lib/Buzz/Client/MultiCurl.php
@@ -2,10 +2,13 @@
 
 namespace Buzz\Client;
 
+use Buzz\Converter\RequestConverter;
+use Buzz\Converter\ResponseConverter;
 use Buzz\Exception\RequestException;
 use Buzz\Message\MessageInterface;
 use Buzz\Message\RequestInterface;
 use Buzz\Exception\ClientException;
+use Psr\Http\Message\RequestInterface as PSR7RequestInterface;
 
 class MultiCurl extends AbstractCurl implements BatchClientInterface
 {
@@ -31,10 +34,20 @@ class MultiCurl extends AbstractCurl implements BatchClientInterface
      * @param RequestInterface $request  A request object
      * @param MessageInterface $response A response object
      * @param array            $options  An array of options
+     *
+     * @deprecated Will be removed in 1.0. Use sendRequest instead.
      */
     public function send(RequestInterface $request, MessageInterface $response, array $options = array())
     {
+        @trigger_error('MultiCurl::send() is deprecated. Use MultiCurl::sendRequest instead.', E_USER_DEPRECATED);
+
+        $request = RequestConverter::psr7($request);
         $this->queue[] = array($request, $response, $options);
+    }
+
+    public function sendRequest(PSR7RequestInterface $request, $options = [])
+    {
+        $this->queue[] = array($request, null, $options);
     }
 
     public function count()
@@ -91,7 +104,8 @@ class MultiCurl extends AbstractCurl implements BatchClientInterface
 
                 // populate the response object
                 if (CURLE_OK === $done['result']) {
-                    static::populateResponse($curl, curl_multi_getcontent($curl), $response);
+                    $psr7Response = $this->createResponse($curl, curl_multi_getcontent($curl));
+                    ResponseConverter::copy(ResponseConverter::buzz($psr7Response), $response);
                 } else if (!isset($e)) {
                     $errorMsg = curl_error($curl);
                     $errorNo  = curl_errno($curl);
@@ -107,7 +121,8 @@ class MultiCurl extends AbstractCurl implements BatchClientInterface
 
                 // callback
                 if (isset($options['callback'])) {
-                    call_user_func($options['callback'], $this, $request, $response, $options, $done['result']);
+                    $returnResponse = isset($options['psr7_response']) && $options['psr7_response'] === true ? $psr7Response : $response;
+                    call_user_func($options['callback'], $this, $request, $returnResponse, $options, $done['result']);
                 }
             }
         }