+++ test/Buzz/Test/Client/FunctionalTest.php
@@ -18,18 +18,19 @@ class FunctionalTest extends TestCase
 {
     protected function setUp()
     {
-        if (!isset($_SERVER['TEST_SERVER'])) {
+        if (!isset($_SERVER['BUZZ_TEST_SERVER'])) {
             $this->markTestSkipped('The test server is not configured.');
         }
     }
 
     /**
      * @dataProvider provideClientAndMethod
+     * @group legacy
      */
     public function testRequestMethods($client, $method)
     {
         $request = new Request($method);
-        $request->fromUrl($_SERVER['TEST_SERVER']);
+        $request->fromUrl($_SERVER['BUZZ_TEST_SERVER']);
         $request->setContent('test');
         $response = $this->send($client, $request);
 
@@ -40,11 +41,12 @@ class FunctionalTest extends TestCase
 
     /**
      * @dataProvider provideClient
+     * @group legacy
      */
     public function testGetContentType($client)
     {
         $request = new Request();
-        $request->fromUrl($_SERVER['TEST_SERVER']);
+        $request->fromUrl($_SERVER['BUZZ_TEST_SERVER']);
         $response = $this->send($client, $request);
 
         $data = json_decode($response->getContent(), true);
@@ -54,11 +56,12 @@ class FunctionalTest extends TestCase
 
     /**
      * @dataProvider provideClient
+     * @group legacy
      */
     public function testFormPost($client)
     {
         $request = new FormRequest();
-        $request->fromUrl($_SERVER['TEST_SERVER']);
+        $request->fromUrl($_SERVER['BUZZ_TEST_SERVER']);
         $request->setField('company[name]', 'Google');
         $response = $this->send($client, $request);
 
@@ -70,11 +73,12 @@ class FunctionalTest extends TestCase
 
     /**
      * @dataProvider provideClient
+     * @group legacy
      */
     public function testFormGet($client)
     {
         $request = new FormRequest(FormRequest::METHOD_GET);
-        $request->fromUrl($_SERVER['TEST_SERVER']);
+        $request->fromUrl($_SERVER['BUZZ_TEST_SERVER']);
         $request->setField('search[query]', 'cats');
         $response = $this->send($client, $request);
 
@@ -86,11 +90,12 @@ class FunctionalTest extends TestCase
 
     /**
      * @dataProvider provideClientAndUpload
+     * @group legacy
      */
     public function testFileUpload($client, $upload)
     {
         $request = new FormRequest();
-        $request->fromUrl($_SERVER['TEST_SERVER']);
+        $request->fromUrl($_SERVER['BUZZ_TEST_SERVER']);
         $request->setField('company[name]', 'Google');
         $request->setField('company[logo]', $upload);
         $response = $this->send($client, $request);
@@ -104,11 +109,12 @@ class FunctionalTest extends TestCase
 
     /**
      * @dataProvider provideClient
+     * @group legacy
      */
     public function testJsonPayload($client)
     {
         $request = new Request(RequestInterface::METHOD_POST);
-        $request->fromUrl($_SERVER['TEST_SERVER']);
+        $request->fromUrl($_SERVER['BUZZ_TEST_SERVER']);
         $request->addHeader('Content-Type: application/json');
         $request->setContent(json_encode(array('foo' => 'bar')));
         $response = $this->send($client, $request);
@@ -121,12 +127,13 @@ class FunctionalTest extends TestCase
 
     /**
      * @dataProvider provideClient
+     * @group legacy
      */
     public function testConsecutiveRequests($client)
     {
         // request 1
         $request = new Request(RequestInterface::METHOD_PUT);
-        $request->fromUrl($_SERVER['TEST_SERVER']);
+        $request->fromUrl($_SERVER['BUZZ_TEST_SERVER']);
         $request->addHeader('Content-Type: application/json');
         $request->setContent(json_encode(array('foo' => 'bar')));
         $response = $this->send($client, $request);
@@ -139,7 +146,7 @@ class FunctionalTest extends TestCase
 
         // request 2
         $request = new Request(RequestInterface::METHOD_GET);
-        $request->fromUrl($_SERVER['TEST_SERVER']);
+        $request->fromUrl($_SERVER['BUZZ_TEST_SERVER']);
         $response = $this->send($client, $request);
 
         $data = json_decode($response->getContent(), true);
@@ -150,11 +157,12 @@ class FunctionalTest extends TestCase
 
     /**
      * @dataProvider provideClient
+     * @group legacy
      */
     public function testPlus($client)
     {
         $request = new FormRequest();
-        $request->fromUrl($_SERVER['TEST_SERVER']);
+        $request->fromUrl($_SERVER['BUZZ_TEST_SERVER']);
         $request->setField('math', '1+1=2');
         $response = $this->send($client, $request);
 
@@ -166,11 +174,12 @@ class FunctionalTest extends TestCase
 
     /**
      * @dataProvider provideClient
+     * @group legacy
      */
     public function testRedirectedResponse($client)
     {
         $request = new Request();
-        $request->fromUrl($_SERVER['TEST_SERVER'].'?redirect_to='.$_SERVER['TEST_SERVER']);
+        $request->fromUrl($_SERVER['BUZZ_TEST_SERVER'].'?redirect_to='.$_SERVER['BUZZ_TEST_SERVER']);
         $response = $this->send($client, $request);
 
         $headers = $response->getHeaders();
@@ -179,6 +188,7 @@ class FunctionalTest extends TestCase
 
     /**
      * @dataProvider provideClient
+     * @group legacy
      */
     public function testProxy($client)
     {
@@ -189,7 +199,7 @@ class FunctionalTest extends TestCase
         $client->setProxy($_SERVER['TEST_PROXY']);
 
         $request = new Request();
-        $request->fromUrl($_SERVER['TEST_SERVER']);
+        $request->fromUrl($_SERVER['BUZZ_TEST_SERVER']);
         $response = $this->send($client, $request);
 
         $data = json_decode($response->getContent(), true);
@@ -199,15 +209,19 @@ class FunctionalTest extends TestCase
     /**
      * @expectedException RuntimeException
      * @expectedExceptionMessage not supported or disabled in libcurl
+     * @group legacy
      */
     public function testRedirectedToForbiddenProtocol()
     {
         $client = new Curl();
         $request = new Request();
-        $request->fromUrl($_SERVER['TEST_SERVER'].'?redirect_to=pop3://localhost/');
+        $request->fromUrl($_SERVER['BUZZ_TEST_SERVER'].'?redirect_to=pop3://localhost/');
         $response = $this->send($client, $request);
     }
 
+    /**
+     * @group legacy
+     */
     public function testMultiCurlExecutesRequestsConcurently()
     {
         $client = new MultiCurl();
@@ -220,7 +234,7 @@ class FunctionalTest extends TestCase
 
         for ($i = 3; $i > 0; $i--) {
             $request = new Request();
-            $request->fromUrl($_SERVER['TEST_SERVER'].'?delay='.$i);
+            $request->fromUrl($_SERVER['BUZZ_TEST_SERVER'].'?delay='.$i);
             $client->send($request, new Response(), array('callback' => $callback));
         }